Rewrite of cluster support
