+++ src/Proto.Cluster.TestProvider/QueryOptions.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Proto.Cluster.Testing
 {
-    internal class QueryOptions
+    public class QueryOptions
     {
         public ulong WaitIndex { get; set; }
         public TimeSpan WaitTime { get; set; }
